Dontprint: Remove HTTP calls to zotero repo
The zotero connector's don't include any translators. Instead,
they download the translators by a localhost HTTP call from
zotero standalone. If zotero standalone is not available, they
fall back to downloading from a public server hosted by zotero.
These files are only chached for the duration of the Browser
session since the Zotero Conector apparently assumes that the
Standalone Zotero is usually available anyway.

For Dontprint, this meant that, since Zotero Standalone is usually
not available, Dontprint on Chrome tries to download all required
translators for each Browser session, which would be extremely
wasteful. This commit removes the HTTP calls to the public Zotero
server and replaces them with "downloads" of translators that are
included in the Dontprint extension bundle.